requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
numpy>=1.24.0
matplotlib>=3.7.0
scipy>=1.11.0
python-dotenv==1.0.0
aiofiles==23.2.1
